fix(api-gateway): map reconcile_attempts to Java long
CockroachDB's INT is INT8 (BIGINT), so the V8 column is BIGINT on the
wire. The entity field was Java int → JDBC INTEGER, which Hibernate
schema validation rejects on startup, crashing the gateway. Widen the
field to long to match storage, consistent with gateway_ts_ms.

+++ api-gateway/src/main/java/com/onlinejudge/gateway/model/Submission.java
@@ -59,5 +59,5 @@ public class Submission {
      * republishing and routes the submission to the DLQ (status FAILED).
      */
     @Column(name = "reconcile_attempts", nullable = false)
-    private int reconcileAttempts = 0;
+    private long reconcileAttempts = 0;
 }
